ENH: Created two Mixin classes. One for widget, one for logic (issue #12) - this could probably useful for further projects
- Calling storescp after selection of intraop directory which might fail, since Slicer has no sudo access (issue #5)
