+++ Utils/helpers.py
@@ -1,0 +1,75 @@
+import logging, qt, ctk, os, slicer
+
+
+class ModuleWidgetMixin(object):
+
+  @staticmethod
+  def makeProgressIndicator(maxVal, initialValue=0):
+    progressIndicator = qt.QProgressDialog()
+    progressIndicator.minimumDuration = 0
+    progressIndicator.modal = True
+    progressIndicator.setMaximum(maxVal)
+    progressIndicator.setValue(initialValue)
+    progressIndicator.setWindowTitle("Processing...")
+    progressIndicator.show()
+    progressIndicator.autoClose = False
+    return progressIndicator
+
+  @staticmethod
+  def confirmDialog(message, title='SliceTracker'):
+    result = qt.QMessageBox.question(slicer.util.mainWindow(), title, message,
+                                     qt.QMessageBox.Ok | qt.QMessageBox.Cancel)
+    return result == qt.QMessageBox.Ok
+
+  @staticmethod
+  def notificationDialog(message, title='SliceTracker'):
+    return qt.QMessageBox.information(slicer.util.mainWindow(), title, message)
+
+  @staticmethod
+  def yesNoDialog(message, title='SliceTracker'):
+    result = qt.QMessageBox.question(slicer.util.mainWindow(), title, message,
+                                     qt.QMessageBox.Yes | qt.QMessageBox.No)
+    return result == qt.QMessageBox.Yes
+
+  @staticmethod
+  def warningDialog(message, title='SliceTracker'):
+    return qt.QMessageBox.warning(slicer.util.mainWindow(), title, message)
+
+
+class ModuleLogicMixin(object):
+
+  @staticmethod
+  def createDirectory(directory, message=None):
+    if message:
+      logging.debug(message)
+    try:
+      os.makedirs(directory)
+    except OSError:
+      logging.debug('Failed to create the following directory: ' + directory)
+
+  @staticmethod
+  def getDICOMValue(currentFile, tag, fallback=None):
+    db = slicer.dicomDatabase
+    try:
+      value = db.fileValue(currentFile, tag)
+    except RuntimeError:
+      logging.info("There are problems with accessing DICOM values from file %s" % currentFile)
+      value = fallback
+    return value
+
+  @staticmethod
+  def getFileList(directory):
+    return [f for f in os.listdir(directory) if ".DS_Store" not in f]
+
+  @staticmethod
+  def importStudy(dicomDataDir):
+    indexer = ctk.ctkDICOMIndexer()
+    indexer.addDirectory(slicer.dicomDatabase, dicomDataDir)
+    indexer.waitForImportFinished()
+
+  @staticmethod
+  def createScalarVolumeNode(name):
+    volume = slicer.vtkMRMLScalarVolumeNode()
+    volume.SetName(name)
+    slicer.mrmlScene.AddNode(volume)
+    return volume
